[mod] #93 네트워크 연결 확인 앱 전역으로 변경
