+++ app/src/main/java/sopt/motivoo/presentation/SplashFragment.kt
@@ -1,11 +1,7 @@
 package sopt.motivoo.presentation
 
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.NetworkCapabilities
 import android.os.Bundle
 import android.view.View
-import androidx.appcompat.app.AlertDialog
 import androidx.lifecycle.lifecycleScope
 import androidx.navigation.NavOptions
 import androidx.navigation.fragment.findNavController
@@ -27,24 +23,7 @@ class SplashFragment : BindingFragment<FragmentSplashBinding>(R.layout.fragment_
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
-        if (checkNetworkState()) {
-            showSplash()
-        } else {
-            showNetworkErrorDialog()
-        }
-    }
-
-    // TODO-l2zh 다이얼로그 나오면 수정
-    private fun showNetworkErrorDialog() {
-        AlertDialog.Builder(requireContext()).apply {
-            setTitle("네트워크 오류")
-            setMessage("네트워크 연결을 확인해주세요.")
-            setPositiveButton("확인") { _, _ ->
-                activity?.finish()
-            }
-            setCancelable(false)
-            create().show()
-        }
+        showSplash()
     }
 
     private fun showSplash() {
@@ -62,15 +41,6 @@ class SplashFragment : BindingFragment<FragmentSplashBinding>(R.layout.fragment_
             .build()
 
         navController.navigate(R.id.action_splashFragment_to_permissionFragment, null, navOptions)
-    }
-
-    private fun checkNetworkState(): Boolean {
-        val connectivityManager =
-            requireContext().getSystemService(Context.CONNECTIVITY_SERVICE) as ConnectivityManager
-        val network = connectivityManager.activeNetwork ?: return false
-        val actNetwork = connectivityManager.getNetworkCapabilities(network) ?: return false
-        return actNetwork.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR) ||
-                actNetwork.hasTransport(NetworkCapabilities.TRANSPORT_WIFI)
     }
 
     companion object {
